requirements.txt
Flask==3.1.2
flask_mail==0.10.0
pymongo==4.15.4
python-dotenv==1.2.1
